Final version lesson5 HomeWork

+++ HomeWorkApp/src/main/java/school/lesson5/task2_1/Fruit.java
@@ -12,7 +12,7 @@ public abstract class Fruit {
         return weight;
     }
 
-    public void setWeight(float weight) {
+    public void setWeight(double weight) {
         this.weight = weight;
     }
